+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
  - Disallow an empty string for a `StoreKey`
+ - `ArrayBuilder` now validates additional fields
 
 ## [0.10.0] - 2024-01-17